+++ python/ry/__init__.py
@@ -7,7 +7,6 @@ from ry import ryo3
 from ry.ryo3 import (
     JSON,
     URL,
-    AsyncClient,
     Date,
     DateDifference,
     DateTime,
@@ -20,6 +19,7 @@ from ry.ryo3 import (
     Glob,
     GlobSet,
     Globster,
+    HttpClient,
     Instant,
     Offset,
     Regex,
@@ -107,7 +107,6 @@ from ry.ryo3 import (
 __all__ = (
     "JSON",
     "URL",
-    "AsyncClient",
     "Date",
     "DateDifference",
     "DateTime",
@@ -120,6 +119,7 @@ __all__ = (
     "Glob",
     "GlobSet",
     "Globster",
+    "HttpClient",
     "Instant",
     "Offset",
     "Regex",